pyproject.toml
[project]
name = "api-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.124.0",
    "ollama>=0.6.1",
    "passlib[argon2,bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
